+++ spec/whois/record/parser/responses/whois.fi/status_available_spec.rb
@@ -21,6 +21,21 @@ describe Whois::Record::Parser::WhoisFi, "status_available.expected" do
     described_class.new(part)
   end
 
+  describe "#disclaimer" do
+    it do
+      subject.disclaimer.should == "More information is available at https://domain.fi/\nCopyright (c) Finnish Communications Regulatory Authority"
+    end
+  end
+  describe "#domain" do
+    it do
+      subject.domain.should == nil
+    end
+  end
+  describe "#domain_id" do
+    it do
+      lambda { subject.domain_id }.should raise_error(Whois::AttributeNotSupported)
+    end
+  end
   describe "#status" do
     it do
       subject.status.should == :available
@@ -43,12 +58,33 @@ describe Whois::Record::Parser::WhoisFi, "status_available.expected" do
   end
   describe "#updated_on" do
     it do
-      lambda { subject.updated_on }.should raise_error(Whois::AttributeNotSupported)
+      subject.updated_on.should == nil
     end
   end
   describe "#expires_on" do
     it do
       subject.expires_on.should == nil
+    end
+  end
+  describe "#registrar" do
+    it do
+      lambda { subject.registrar }.should raise_error(Whois::AttributeNotSupported)
+    end
+  end
+  describe "#registrant_contacts" do
+    it do
+      subject.registrant_contacts.should be_a(Array)
+      subject.registrant_contacts.should == []
+    end
+  end
+  describe "#admin_contacts" do
+    it do
+      lambda { subject.admin_contacts }.should raise_error(Whois::AttributeNotSupported)
+    end
+  end
+  describe "#technical_contacts" do
+    it do
+      lambda { subject.technical_contacts }.should raise_error(Whois::AttributeNotSupported)
     end
   end
   describe "#nameservers" do