+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - NEW: Added full whois.srs.net.nz parser.
 
+- NEW: Added full whois.fi parser.
+
 - CHANGED: Updated whois.nic.ac, whois.nic.io, and whois.nic.sh parsers to the new response format (GH-238).
 
 - CHANGED: Updated whois1.nic.bi parser to the new response format.